+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Allow to drain on-chain wallet by sending amount `0`.
+
 ## [1.4.4] - 2023-10-28
 
 - Improve collab revert